+++ CHANGELOG.md
@@ -10,6 +10,7 @@ See also <https://github.com/xipki/xipki/releases>
   - Extend the command csr-p11 and csr-p12 to generate CSR with complex subject and extensions
   - In the extension CertificatePolicies, the OID for User Notice is not correct. This has been fixed.
   - Add the management of the certificate of parenet CAs for given CA
+  - Extension AuthorityKeyIdentifier embeds both KeyIdentifier and (authorityCertIssuer, authorityCertSerialNumber) in case of incorrect configuration. However exactly one of them is allowed. This has been fixed.
 
 ## 5.1.0
   - Release date: Mar 17, 2019